Split framework module into common/json/worker module
