+++ framework/src/test/scala/skinny/controller/ParamsSpec.scala
@@ -1,7 +1,7 @@
 package skinny.controller
 
-import org.scalatra.test.scalatest._
-import skinny._, controller._
+import org.scalatra.test.scalatest.ScalatraFlatSpec
+import skinny._
 
 class ParamsSpec extends ScalatraFlatSpec {
 
@@ -20,7 +20,8 @@ class ParamsSpec extends ScalatraFlatSpec {
         ("year", "month", "day", "hour", "minute", "second"), "datetime").getAs[String]("datetime").orNull
     }
     def datetime2 = {
-      Params(params).withDateTime("datetime").getAs[String]("datetime").orNull
+      Params(params).
+        withDateTime("datetime").getAs[String]("datetime").orNull
     }
     def datetime3 = {
       Params(params).withDateTime(
@@ -70,64 +71,6 @@ class ParamsSpec extends ScalatraFlatSpec {
     }
     post("/time2", "time_hour" -> "1", "time_minute" -> "23", "time_second" -> "4") {
       body should equal("1970-01-01 01:23:04")
-    }
-  }
-
-  it should "work with nullable values" in {
-    // Some(null) must be avoided
-    {
-      val params = Params(Map("something" -> "foo"))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(Some("foo"))
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
-    }
-    {
-      val params = Params(Map("something" -> null))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(None)
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
-    }
-    {
-      val params = Params(Map("something" -> Some("foo")))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(Some("foo"))
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
-    }
-    {
-      val params = Params(Map("something" -> Some(Some("foo"))))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(Some(Some("foo")))
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
-    }
-    {
-      val params = Params(Map("something" -> Some(null)))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(None)
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
-    }
-    {
-      val params = Params(Map("something" -> None))
-
-      params.something.exists(_ == null) should equal(false)
-      params.something should equal(None)
-
-      params.other.exists(_ == null) should equal(false)
-      params.other.isDefined should equal(false)
     }
   }
