+++ worker/build.sbt
@@ -1,0 +1,6 @@
+scalariformSettings
+
+net.virtualvoid.sbt.graph.Plugin.graphSettings
+
+instrumentSettings
+
