+++ worker/src/main/scala/skinny/worker/SkinnyWorker.scala
