+++ worker/src/test/resources/logback.xml
@@ -1,0 +1,17 @@
+<configuration>
+    <appender name="default" class="ch.qos.logback.core.FileAppender">
+        <file>logs/console.log</file>
+        <encoder>
+          <pattern>%date %level [%thread] %logger{64} %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <logger name="org.eclipse.jetty" level="warn"/>
+    <logger name="org.fusesource.scalate" level="warn"/>
+    <logger name="org.scalatra.scalate.ScalateSupport##anonfun#createTemplateEngine#2##anon#1.SourceMap" level="warn"/>
+    <logger name="org.scalatra.scalate.ScalateSupport##anonfun#createTemplateEngine#2##anon#2.SourceMap" level="warn"/>
+
+    <root level="debug">
+        <appender-ref ref="default"/>
+    </root>
+</configuration>
