+++ worker/src/test/scala/skinny/worker/SkinnyWorkerSpec.scala
